+++ copu1.txt
