+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "main": "index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "node src/index.js --username=Rarity",
     "start:n": "node src/index.js --username=Rarity"
   },
   "repository": {
