+++ src/consts.js
@@ -1,5 +1,7 @@
+import { argv } from 'node:process';
+
 const parseArgs = () => {
-    const args = process.argv.slice(2).toString().trim();
+    const args = argv.slice(2).toString().trim();
     const userName = args.startsWith('--username') ? args.split('=').pop() : "Anonimus";
     return userName;
 };
